Create com_bus.rs for com_rx, com_tx, etc.
This moves com_rx, com_tx, and com_int_handler from sw/src/main.rs
into their own module at sw/src/com_bus.rs.
